# src/main/resources/static/api-docs.yml
swagger: "2.0"
info:
  description: >-
    Especificaciones técnicas para consumo de la api DigitalBanking
  version: 1.0.1
  title: Rest Api DigitalBanking
  termsOfService: http://swagger.io/terms/
  contact:
    email: soporte@digitalbanking.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:8080
basePath: /api/v1
...
schemes:
  - https
  - http
paths:
...
  /imagen/buscarTodos:
    post:
      tags:
        - Imagenes
      summary: Obtiene todas las imagenes
      description: ""
      operationId: buscarImagenes
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/PagineableDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            type: array
            items:
              $ref: "#/definitions/ImagenDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      deprecated: false
  /imagen/upload:
    post:
      tags:
        - Imagenes
      summary: Carga una imagen en el bucket s3
      description: ""
      operationId: cargarImagen
      produces:
        - application/json
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: imagen
          type: file
          description: imagen
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/ImagenDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /imagen/{id}:
    get:
      tags:
        - Imagenes
      summary: Obtiene una imagen por id
      description: ""
      operationId: buscarImagenPorId
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          required: true
          name: id
          description: id
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/ImagenDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      deprecated: false
  /imagen/eliminar/{id}:
    delete:
      tags:
        - Imagenes
      summary: Elimina una imagen
      description: ""
      operationId: eliminarImagen
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          name: id
          required: true
          description: id
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /imagen/desactivar/{key}:
    delete:
      tags:
        - Imagenes
      summary: Desactiva una imagen
      description: ""
      operationId: desactivarImagen
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          name: key
          required: true
          description: key
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /imagen:
    post:
      tags:
        - Imagenes
      summary: Crea una imagen nueva
      description: ""
      operationId: crearImagen
      produces:
        - application/json
      parameters:
        - in: body
          name: id
          description: request
          schema:
            $ref: "#/definitions/ImagenDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/ImagenDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
    put:
      tags:
        - Imagenes
      summary: Actualizar una imagen
      description: ""
      operationId: actualizarImagen
      produces:
        - application/json
      parameters:
        - in: body
          name: id
          description: request
          schema:
            $ref: "#/definitions/ImagenDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/ImagenDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
  /rol/buscarTodos:
    post:
      tags:
        - Roles
      summary: Obtiene todos las roles
      description: ""
      operationId: buscarRoles
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: request
          schema:
            $ref: "#/definitions/PagineableDto"
      responses:
        "200":
          description: Operacion exitosa
          schema:
            type: array
            items:
              $ref: "#/definitions/RolDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
...
  /rol/{id}:
    get:
      tags:
        - Roles
      summary: Obtiene un rol por id
      description: ""
      operationId: buscarRolPorId
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          required: true
          name: id
          description: id
      responses:
        "200":
          description: Operacion exitosa
          schema:
            $ref: "#/definitions/RolDto"
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
    delete:
      tags:
        - Roles
      summary: Elimina un rol
      description: ""
      operationId: eliminarRol
      produces:
        - application/json
      parameters:
        - in: path
          type: string
          name: id
          required: true
          description: id
      responses:
        "200":
          description: Operacion exitosa
        "400":
          description: Request incorrecto
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: No autorizado
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: Acceso prohibido
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: Error inesperado
          schema:
            $ref: "#/definitions/Error"
      security:
        - APIKeyHeader: [ ]
      deprecated: false
...
securityDefinitions:
  APIKeyHeader:
    type: apiKey
    in: header
    name: Authorization
definitions:
  PagineableDto:
    type: object
    properties:
      pagina:
        type: integer
        format: int32
      cantidad:
        type: integer
        format: int32
...
  ImagenDto:
    type: object
    properties:
      idImagen:
        type: integer
        format: int32
      keyImagen:
        type: string
      urlImagen:
        type: string
      indicadorHabilitado:
        type: boolean
  RolDto:
    type: object
    properties:
      idRol:
        type: integer
        format: int32
      codigo:
        type: string
      nombre:
        type: string
      indicadorHabilitado:
        type: boolean
...
  Error:
    type: object
    properties:
      timestamp:
        type: integer
        format: int32
      status:
        type: integer
        format: int32
      error:
        type: string
      path:
        type: string
...